Get dsSrc data from google sheet

The dsSrc dataset is maintained by Julie and contains most of the relevant onboarding information.

Usage

dssrc_dataframe()

Examples

dssrc = dssrc_dataframe()
#> ! Using an auto-discovered, cached token.
#>   To suppress this message, modify your code or options to clearly consent to
#>   the use of a cached token.
#>   See gargle's "Non-interactive auth" vignette for more details:
#>   <https://gargle.r-lib.org/articles/non-interactive-auth.html>
#> ℹ The googlesheets4 package is using a cached token for seandavi@gmail.com.
#> ✔ Reading from Directory-SRC-PROD PRIVATE.
#> ✔ Range ''dsSrc'!2:10000000'.
colnames(dssrc)
#>  [1] "orcid"             "fullName"          "preferredFullName"
#>  [4] "EndDate"           "pronouns-all"      "email"            
#>  [7] "driveEmail-all"    "pronunciation"     "githubHandle"     
#> [10] "assistantEmail"    "bio"               "centerOrProgram"  
#> [13] "activities-all"    "expertise-all"     "degrees-all"      
#> [16] "Affils-all"        "isOffboarded"      "allRoles"         
#> [19] "allGroups"         "searchString"      "primaryAffilName" 
dplyr::glimpse(dssrc)
#> Rows: 275
#> Columns: 21
#> $ orcid             <chr> "0000-0001-5011-4282", "0000-0001-5015-4428", "0000-…
#> $ fullName          <chr> "Brooke A. Armfield, PhD", "Christopher R. Kinsinger…
#> $ preferredFullName <chr> "Brooke Armfield", "Chris Kinsinger", "Ben Cordier",…
#> $ EndDate           <chr> "Wed Nov 16 2022 17:54:12 GMT-0800 (Pacific Standard…
#> $ `pronouns-all`    <chr> "she/her", "he/him", "he/him", "he/him", "she/her", …
#> $ email             <chr> "barmfield@ufl.edu", "chris.kinsinger@nih.gov", "cor…
#> $ `driveEmail-all`  <chr> "barmfield@ufl.edu", "kinsinge2000@gmail.com", "bena…
#> $ pronunciation     <chr> NA, "Chris Kinsinger, pronounced \"Kin - singer\"; n…
#> $ githubHandle      <chr> NA, NA, "greenstick", NA, NA, NA, NA, NA, NA, "nshef…
#> $ assistantEmail    <chr> NA, NA, NA, NA, NA, "sstanl20@jhmi.edu", NA, NA, NA,…
#> $ bio               <chr> NA, NA, "https://benjamincordier.com", "www.berkustu…
#> $ centerOrProgram   <chr> "dgps", "dgps", "dgps", "cores", "dgps", "cores", "d…
#> $ `activities-all`  <chr> "CHoRUS DGP", "CHoRUS DGP", "AI-READi DGP", "Ethics …
#> $ `expertise-all`   <chr> "R, Other Standards, Other Teaming, Other Ethics, Ot…
#> $ `degrees-all`     <chr> "PhD", "PhD", "BA, MS, PhD", "PhD", "PhD", "MD, DrPH…
#> $ `Affils-all`      <chr> "University of Florida", "National Institutes of Hea…
#> $ isOffboarded      <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FAL…
#> $ allRoles          <chr> "CHoRUS DGP - member; CHoRUS Ethics Module - PM; CHo…
#> $ allGroups         <chr> "chorus-dgp@b2ai.org;chorus-ethics-module@b2ai.org;c…
#> $ searchString      <chr> "wed nov 16 2022 17:54:12 gmt-0800 (pacific standard…
#> $ primaryAffilName  <chr> "University of Florida", "National Institutes of Hea…