Get a ORCID work detail

The ORCID API has the concept of a work with its own dedicated API endpoint. This function calls that endpoint and returns a single data.frame row representing that work.

Usage

orcid_work_detail(orcid, put_code)

Arguments

orcid

character(1) an orcid is required

put_code

integer(1) the put-code, an ORCID API internal identifier for a work

Details

Additional functionality is available if the work record has a bibtex entry. Using the rbibutils package, one can convert the citation to bibtex, latex, citations of various styles, and even HTML.

See also

Examples

orcid = '0000-0001-5643-4068'
put_code = 89488980

df = orcid_work_detail(orcid, put_code)
df
#>                 orcid put_code          created_at          updated_at
#> 1 0000-0001-5643-4068 89488980 2021-02-25 18:56:13 2022-05-31 22:54:57
#>   citation_type
#> 1        bibtex
#>                                                                                                                                                                                                                                                                      citation_value
#> 1 @article{PMID:31995185,\n\ttitle = {Refgenie: a reference genome resource manager.},\n\tauthor = {Stolarczyk M and Reuter VP and Smith JP and Magee NE and Sheffield NC},\n\tdoi = {10.1093/gigascience/giz149},\n\tjournal = {GigaScience},\n\tmonth = {02},\n\tyear = {2020}\n}
#>         work_type contributors                                        url
#> 1 journal-article           NA http://europepmc.org/abstract/med/31995185
#>                          doi     pmid        pmc abstract
#> 1 10.1093/gigascience/giz149 31995185 PMC6988606       NA
colnames(df)
#>  [1] "orcid"          "put_code"       "created_at"     "updated_at"    
#>  [5] "citation_type"  "citation_value" "work_type"      "contributors"  
#>  [9] "url"            "doi"            "pmid"           "pmc"           
#> [13] "abstract"      

# the bibtex entry, if available,
# can be used for additional manipulation
if(require('rbibutils')) {
   rref = rbibutils::charToBib(df$citation_value)
   print(rref)

   # bibtex
   print(rref, style = "Bibtex")

   # simple citation
   print(rref, style = "citation")

   # html
   print(rref, style = "html")

   # latex
   print(rref, style = "latex")

   # r citation
   print(rref, style = "R")
}
#> Loading required package: rbibutils
#> M S, VP R, JP S, NE M, NC S (2020). “Refgenie: a reference genome
#> resource manager.” _GigaScience_. doi:10.1093/gigascience/giz149
#> <https://doi.org/10.1093/gigascience/giz149>,
#> <https://doi.org/10.1093/gigascience/giz149>.
#> @Article{PMID:31995185,
#>   author = {Stolarczyk M and Reuter VP and Smith JP and Magee NE and Sheffield NC},
#>   title = {Refgenie: a reference genome resource manager.},
#>   journal = {GigaScience},
#>   year = {2020},
#>   month = {Feb},
#>   doi = {10.1093/gigascience/giz149},
#>   url = {https://doi.org/10.1093/gigascience/giz149},
#> }
#> 
#> M S, VP R, JP S, NE M, NC S (2020). “Refgenie: a reference genome
#> resource manager.” _GigaScience_. doi:10.1093/gigascience/giz149
#> <https://doi.org/10.1093/gigascience/giz149>,
#> <https://doi.org/10.1093/gigascience/giz149>.
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Article{PMID:31995185,
#>     author = {Stolarczyk M and Reuter VP and Smith JP and Magee NE and Sheffield NC},
#>     title = {Refgenie: a reference genome resource manager.},
#>     journal = {GigaScience},
#>     year = {2020},
#>     month = {Feb},
#>     doi = {10.1093/gigascience/giz149},
#>     url = {https://doi.org/10.1093/gigascience/giz149},
#>   }
#> 
#> 
#> 
#> <p>M S, VP R, JP S, NE M, NC S (2020).
#> &ldquo;Refgenie: a reference genome resource manager.&rdquo;
#> <em>GigaScience</em>.
#> <a href="https://doi.org/10.1093/gigascience/giz149">doi:10.1093/gigascience/giz149</a>, <a href="https://doi.org/10.1093/gigascience/giz149">https://doi.org/10.1093/gigascience/giz149</a>. 
#> </p>
#> M S, VP R, JP S, NE M, NC S (2020).
#> ``Refgenie: a reference genome resource manager.''
#> \emph{GigaScience}.
#> \Rhref{https://doi.org/10.1093/gigascience/giz149}{doi:10.1093\slash{}gigascience\slash{}giz149}, \url{https://doi.org/10.1093/gigascience/giz149}.
#> bibentry(bibtype = "Article",
#>          key = "PMID:31995185",
#>          author = c(person(given = "Stolarczyk",
#>                            family = "M"),
#>                     person(given = "Reuter",
#>                            family = "VP"),
#>                     person(given = "Smith",
#>                            family = "JP"),
#>                     person(given = "Magee",
#>                            family = "NE"),
#>                     person(given = "Sheffield",
#>                            family = "NC")),
#>          title = "Refgenie: a reference genome resource manager.",
#>          journal = "GigaScience",
#>          year = "2020",
#>          month = "Feb",
#>          doi = "10.1093/gigascience/giz149",
#>          url = "https://doi.org/10.1093/gigascience/giz149")